requirements.txt
nextcord==2.0.0rc1
python-dotenv==0.20.0
requests==2.28.0
types-requests==2.27.30
pymongo==4.1.1
dnspython==2.2.1
psutil==5.9.1
humanfriendly==10.0
mojang==0.2.0
pillow==9.1.1
coloredlogs==15.0.1
aiohttp==3.8.1
